Add files via upload
Added a few plots just to get the folder structure right. This is important if someone wishes to download the software and run it, as the plotting functionalities assume that the right folders can be found where they should be. Apparently empty folders cannot be instantiated for this purpose here.